db_backend/requirements.txt
langchain-gigachat
langgraph
psycopg2-binary
configparser
sqlalchemy
